pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "aid-commander"
version = "4.0.0"
description = "AI-Facilitated Iterative Development with Memory Bank - v4.0"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "AID Development Team", email = "aid@sistronics.com"}
]
requires-python = ">=3.8"
dependencies = [
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "watchdog>=3.0.0",
    "jinja2>=3.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
ai = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "azure-cognitiveservices-language-textanalytics>=5.3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.5",
    "pre-commit>=3.0",
]
ide = [
    "language-server-protocol>=3.16.0",
    "pygls>=1.0.0",
    "jsonrpc-server>=1.2.0",
]
all = [
    "aiohttp>=3.8.0", "aiofiles>=23.0.0", "aiosqlite>=0.19.0", "pydantic>=2.0.0", "rich>=13.0.0",
    "typer>=0.9.0", "cryptography>=41.0.0", "python-dotenv>=1.0.0", "watchdog>=3.0.0", 
    "jinja2>=3.1.0", "numpy>=1.24.0", "scikit-learn>=1.3.0",
    "openai>=1.0.0", "anthropic>=0.3.0", "azure-cognitiveservices-language-textanalytics>=5.3.0",
    "pytest>=7.0", "pytest-asyncio>=0.21.0", "pytest-cov>=4.0", "black>=23.0",
    "flake8>=6.0", "mypy>=1.5", "pre-commit>=3.0",
    "language-server-protocol>=3.16.0", "pygls>=1.0.0", "jsonrpc-server>=1.2.0",
]

[project.scripts]
aid-commander = "aid_commander:main"
aid-memory = "memory_service:cli_main"
aid-quality = "quality_gates:cli_main"
aid-context = "context_engine:cli_main"
aid-conversation = "conversation_manager:cli_main"

[tool.setuptools]
py-modules = [
    "aid_commander", "memory_service", "context_engine", "quality_gates", 
    "conversation_manager", "ai_service", "template_engine", "logger"
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=. --cov-report=html"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true